Microphone max amplitude bug fix

+++ app/src/main/java/com/gmail/btheo95/musicflashlight/resource/SoundMeter.java
@@ -38,7 +38,7 @@ public class SoundMeter {
         }
     }
 
-    public double getAmplitude() {
+    public int getAmplitude() {
         if (mRecorder != null)
             return mRecorder.getMaxAmplitude();
         else
